# Improve ctx.Send and ctx.SendBytes performance