--- a/ctx_test.go
+++ b/ctx_test.go
@@ -1418,3 +1418,29 @@ func Test_Ctx_XHR(t *testing.T) {
 	ctx.Fasthttp.Request.Header.Set(HeaderXRequestedWith, "XMLHttpRequest")
 	utils.AssertEqual(t, true, ctx.XHR())
 }
+
+// go test -v  -run=^$ -bench=Benchmark_Ctx_SendString_For_Reset -benchmem -count=4
+func Benchmark_Ctx_SendString_For_Reset(b *testing.B) {
+	app := New()
+	c := app.AcquireCtx(&fasthttp.RequestCtx{})
+	defer app.ReleaseCtx(c)
+	b.ReportAllocs()
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		c.SendString("")
+	}
+	utils.AssertEqual(b, "", string(c.Fasthttp.Response.Body()))
+}
+
+// go test -v  -run=^$ -bench=Benchmark_Ctx_Reset_Body -benchmem -count=4
+func Benchmark_Ctx_Reset_Body(b *testing.B) {
+	app := New()
+	c := app.AcquireCtx(&fasthttp.RequestCtx{})
+	defer app.ReleaseCtx(c)
+	b.ReportAllocs()
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		c.Fasthttp.Response.ResetBody()
+	}
+	utils.AssertEqual(b, "", string(c.Fasthttp.Response.Body()))
+}