NEW
```go
Benchmark_Ctx_SendString_For_Reset
Benchmark_Ctx_SendString_For_Reset-8    100000000               11.8 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_SendString_For_Reset-8    100000000               11.6 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_SendString_For_Reset-8    100000000               12.0 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_SendString_For_Reset-8    100000000               11.9 ns/op             0 B/op          0 allocs/op

Benchmark_Ctx_Reset_Body
Benchmark_Ctx_Reset_Body-8      169810179                7.03 ns/op            0 B/op          0 allocs/op
Benchmark_Ctx_Reset_Body-8      170163852                7.05 ns/op            0 B/op          0 allocs/op
Benchmark_Ctx_Reset_Body-8      168444130                7.11 ns/op            0 B/op          0 allocs/op
Benchmark_Ctx_Reset_Body-8      168826455                7.25 ns/op            0 B/op          0 allocs/op
```

OLD
```go
Benchmark_Ctx_SendBytes_With_SendString
Benchmark_Ctx_SendBytes_With_SendString-8       84163512                14.7 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_SendBytes_With_SendString-8       86428518                14.7 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_SendBytes_With_SendString-8       84299564                14.6 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_SendBytes_With_SendString-8       83541979                14.6 ns/op             0 B/op          0 allocs/op

Benchmark_Ctx_SendBytes_With_SetBodyRaw
Benchmark_Ctx_SendBytes_With_SetBodyRaw-8       127097104                7.89 ns/op            0 B/op          0 allocs/op
Benchmark_Ctx_SendBytes_With_SetBodyRaw-8       126810813                8.05 ns/op            0 B/op          0 allocs/op
Benchmark_Ctx_SendBytes_With_SetBodyRaw-8       127799650                7.98 ns/op            0 B/op          0 allocs/op
Benchmark_Ctx_SendBytes_With_SetBodyRaw-8       126872679                7.89 ns/op            0 B/op          0 allocs/op

```