@ReneWerner87 2020-07-09
Can you share the benchmark times before your adjustment ?